v0.4.0 Added support for searching multiple paths/possible extensions
This refactors most of grunt-hug. It introduces the "path" and
"extensions" configuration options.

path: Defining an array in path will tell grunt-hug to search
through those paths in order to find the file.

extensions: Defining an array for the extensions options will
tell grunt to guess the path for files when searching for them.
By default it will look for a ".js" match if it doesn't find an
exact match.

For example if you require('jquery'), grunt-hug will first
search for "./jquery", then "./jquery.js". If you had also
defined the path to ["./components"], grunt-hug will continue
to search in "./components/jquery" and "./components/jquery.js"
until it finds a match.

This should be a fairly unstable at the moment, as I haven't
written the unit tests yet. But it seems to work for a complex
project I have right now.

Signed-off-by: Ozan Turgut <oturgut@turn.com>